requirements.txt
click>=7.1.2
termcolor>=1.1.0
setuptools~=51.1.1
ffmpeg-python>=0.2.0